WebServer/requirements.txt
django
channels
channels_redis
channels-redis==2.4.2
djangorestframework
